IP Address to Number
An IPv4 address is a 32-bit number that identifies a device on the internet.

While computers read and write IP addresses as a 32-bit number, we prefer to read them in dotted-decimal notation,
which is basically the number split into 4 chunks of 8 bits, converted to decimal, and delmited by a dot.

In this kata, you will create the function ipToNum that takes an ip address and converts it to a number, 
as well as the function numToIp that takes a number and converts it to an IP address string. Input will always be valid.

Note that the binary octets are not using two's complement (so we can't have negative numbers) even though JavaScript 
numbers do.